Flask/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
black==22.8.0
boto3==1.24.56
botocore==1.27.56
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
factory-boy==3.2.1
Faker==14.1.0
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
krakenio==0.0.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
networkx==2.8.6
numpy==1.23.2
packaging==21.3
password-strength==0.0.3.post2
pathspec==0.10.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
s3transfer==0.6.0
scipy==1.9.0
shortuuid==1.0.9
six==1.16.0
SQLAlchemy==1.4.40
tifffile==2022.8.12
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
